Avoid symlink footgun in prep_stage2.sh

+++ ibd_prod/scripts/prep_stage2.sh
@@ -36,7 +36,11 @@ stage2_vars $tag@$configname
 
 mkdir -p $indir $logdir $trueOutdir
 mkdir -p $(dirname $outdir)
-ln -s $trueOutdir $outdir
+
+# if $outdir already exists, we'll end up making a symlink *inside* it, bad!
+if [ ! -e $outdir ]; then
+    ln -s $trueOutdir $outdir
+fi
 
 conffile=$indir/config.$configname.txt
